backend/ai_agent/requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-openai
langchain-mcp-adapters
mcp
httpx
python-dotenv